refactor tab content into folders
