+++ frontend/src/toolbar/actions/Actions.scss
