+++ frontend/src/toolbar/actions/AllActionsLink.scss
