+++ frontend/src/toolbar/dashboards/AllDashboardsLink.scss
