+++ frontend/src/toolbar/shared/FloatingToolbarHeader.js
