+++ frontend/src/toolbar/shared/InspectElement.js
