+++ frontend/src/toolbar/stats/CurrentPage.scss
