+++ frontend/src/toolbar/stats/currentPageLogic.js
